Forms for creating and updating user validates empty fields
algorithm for creating games array for new users moved to frontend.

A note will hover when the mouse is over an avatar icon in memberlist quicklist.

Fixed a bug in userprofile when a text doesn't show up in games library, if games list is empty

added icons for phone and email in userprofile

alert boxes will pop up when creating or updating user